+++ drivers/block/nd/btt.h
@@ -168,7 +168,7 @@ struct btt {
 	struct dentry *debugfs_dir;
 	struct block_device *backing_dev;
 	struct nd_btt *nd_btt;
-	size_t nlba;
+	u64 nlba;
 	size_t rawsize;
 	u32 lbasize;
 	spinlock_t *lane_lock;
